course251~300
 Awarding XP Game Plan

 XP Reward for Enemies

 Incoming XP Meta Attribute

 Passively Listening for Events

 Sending XP Events

 Showing XP in the HUD

 Level Up Interface Function

 Leveling Up

 Showing Level in the HUD

 Level Up Niagara System

 Level Up HUD Message

 Attribute Points Member Variable

 Showing Attribute Points in the HUD

 Attribute Upgrade Buttons

 Upgrading Attributes

 Top Off Our Fluids

 Attribute Menu Polish

 Spell Menu Design

 Spell Globe Button

 Offensive Spell Tree

 Passive Spell Tree-720P 高清-AVC

 Equipped Spell Row

 Spell Menu Widget

 Spell Description Box

 Spell Menu Button

 Spell Menu Widget Controller

 Constructing the Spell Menu Widget Controller

 Equipped Row Button

 Ability Status and Type

 Showing Abilities in the Spell Tree

 Ability Level Requirement

 Update Ability Statuses

 Updating Status in the Spell Menu

 Show Spell Points

 Selecting Icons

 Deselecting Icons

 Spell Menu Buttons

 Selected Ability

 Spending Spell Points

 Rich Text Blocks

 Spell Descriptions

 FireBolt Description

 Cost and Cooldown in Spell Description

 Self Deselect

 Equipped Spell Row Animations

 Ability Types

 Equipping Abilities

 Updating the Overlay When Equipping Abilities

 Globe Reassigned

 Unbinding Delegates

+++ Source/PJ_AURA/Private/AbilitySystem/AsyncTasks/WaitCooldownChange.cpp
@@ -63,7 +63,7 @@ void UWaitCooldownChange::OnActiveEffectAdded(UAbilitySystemComponent* TargetASC
 		FGameplayEffectQuery Query = FGameplayEffectQuery::MakeQuery_MatchAnyOwningTags(CooldownTag.GetSingleTagContainer());
 		TArray<float> TimesRemaining =  ASC->GetActiveEffectsTimeRemaining(Query);
 		float TimeRemaining = -1.f;
-		for (int32 num : TimesRemaining)
+		for (float num : TimesRemaining)
 		{
 			if (TimeRemaining < 0.f)
 			{
